Remove circular deps from CDI test.
